add: damage observer, revert float to int health

+++ Assets/Scripts/Weapons/DamageStrategy/DamageStrategy.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 
 public abstract class DamageStrategy : MonoBehaviour
 {
-    public abstract float CalculateDamage();
+    public abstract int CalculateDamage();
 }
